requirements.txt
# Core
sqlalchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
httpx>=0.27.0
tenacity>=8.2.0

# CLI & Dashboard
typer[all]>=0.12.0
rich>=13.7.0

# Data processing
pandas>=2.0.0
geopy>=2.4.0

# Scheduling
apscheduler>=3.10.0

# Email
# smtplib is built-in; for HTML email:
email-validator>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: Telegram alerts
# python-telegram-bot>=20.0

# Optional: SMS alerts
# twilio>=8.0.0
